Size is not grandeur, and territory does not make a nation clear q

+++ 0x00-hello_world/6-size.c
@@ -5,15 +5,16 @@
  */
 int main(void)
 {
-char charType;
-int intType;
-long int longintType;
-long long int longlongintType;
-float floatType;
-printf("Size of a char: %lu byte(s)\n", sizeof(charType));
-printf("Size of an int: %lu byte(s)\n", sizeof(intType));
-printf("Size of a long int: %lu byte(s)\n", sizeof(longintType));
-printf("Size of a long long int: %lu byte(s)\n", sizeof(longlongintType));
-printf("Size of a float: %lu byte(s)\n", sizeof(floatType));
+	char charType;
+	int intType;
+	long int longintType;
+	long long int longlongintType;
+	float floatType;
+
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(charType));
+printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(intType));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(longintType));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(longlongintType));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(floatType));
 return (0);
 }
